requirements.windows_gpu.tmp.txt
tensorflow-cpu==2.10.0
tensorflow-directml-plugin
numpy<2
streamlit>=1.31
pandas>=2.0
pillow>=10.0
nibabel>=5.2
scikit-learn>=1.3
scipy>=1.10
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18
